// frontend/src/core/utils/enums/candidates.ts
export enum CandidateStatusEnum {
  applied = "applied",
  resume_scraped = "resume_scraped",
  screened = "screened",
  survey = "survey",
  coding_test = "coding_test",
  interview_lead = "interview_team_lead",
  interview_gm = "interview_general_manager",
  offer = "offer",
  hired = "hired",
  rejected = "rejected",
  all = "all",
}